refactor: change cards to cards folder
