+++ src/views/UniverseView/Components/Cards/ExpectedGrowthRateAndMarginCard.jsx
