+++ src/views/UniverseView/Components/Cards/KPIAveragesCard.jsx
