+++ src/views/UniverseView/UniverseView.jsx
@@ -1,13 +1,13 @@
 import React from 'react'
 import Grid from '@material-ui/core/Grid'
 import Divider from '@material-ui/core/Divider'
-import { CountBySizeCard } from './Components/CountBySizeCard'
-import { KPIAveragesCard } from './Components/KPIAveragesCard'
-import { GrowthAndMarginCard } from './Components/GrowthAndMarginCard'
-import { RevenueAndEbitdaCard } from './Components/RevenueAndEbitdaCard'
-import { ExpectedGrowthRateAndMarginCard } from './Components/ExpectedGrowthRateAndMarginCard'
+import { CountBySizeCard } from './Components/Cards/CountBySizeCard'
+import { KPIAveragesCard } from './Components/Cards/KPIAveragesCard'
+import { GrowthAndMarginCard } from './Components/Cards/GrowthAndMarginCard'
+import { RevenueAndEbitdaCard } from './Components/Cards/RevenueAndEbitdaCard'
+import { ExpectedGrowthRateAndMarginCard } from './Components/Cards/ExpectedGrowthRateAndMarginCard'
+import { Filter } from './Components/Filter/Filter'
 import { RuleGraph } from './Components/RuleGraph'
-import { Filter } from './Components/Filter'
 import useUniverseOverview from '../../hooks/useUniverseOverview'
 import { Information } from './Components/HeaderInformation'
